# services/gateway/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gateway Service API
  version: 1.0.0
  description: REST API exposed by the technoBro gateway service.
servers:
  - url: https://gateway.local
    description: Default environment
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/users/register:
    post:
      tags: [auth, users]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/users/login:
    post:
      tags: [auth, users]
      summary: Login with credentials
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/users/refresh:
    post:
      tags: [auth, users]
      summary: Refresh token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/users/me:
    get:
      security:
        - bearerAuth: []
      tags: [auth, users]
      summary: Get current user info
      description: Returns information about the currently authenticated user
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/requests:
    get:
      security:
        - bearerAuth: []
      tags: [requests]
      summary: List requests
      parameters:
        - in: query
          name: responsible_id
          schema:
            type: string
            format: uuid
          description: Filter by responsible porter
      responses:
        '200':
          description: Requests list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestListResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/requests/{requestId}:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      security:
        - bearerAuth: []
      tags: [requests]
      summary: Get request by ID
      responses:
        '200':
          description: Request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      security:
        - bearerAuth: []
      tags: [requests]
      summary: Partially update a request
      description: >
        Updates one or more fields of a request (status, address, schedule time,
        request text, responsible person). Only provided fields are changed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpdateRequest'
      responses:
        '200':
          description: Updated request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/requests/{requestId}/responsible:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    post:
      security:
        - bearerAuth: []
      tags: [requests]
      summary: Assign responsible porter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignResponsibleRequest'
      responses:
        '200':
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/requests/{requestId}/history:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      security:
        - bearerAuth: []
      tags: [requests]
      summary: List request status history
      responses:
        '200':
          description: History entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestStatusHistoryListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      security:
        - bearerAuth: []
      tags: [requests]
      summary: Append status history entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestStatusUpdateRequest'
      responses:
        '201':
          description: History entry added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestStatusHistoryResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
    RegisterRequest:
      type: object
      required: [username, password, email]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    TokenPair:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    TokenRefreshRequest:
      $ref: '#/components/schemas/TokenPair'
    MeResponse:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, porter]
...
    EquipmentSnapshot:
      type: object
      description: Equipment snapshot embedded in a request
      properties:
        id:
          type: integer
        quantity:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RequestStatus:
      type: string
      enum:
        - canceled
        - pending
        - assigned
        - in_progress
        - completed
        - rejected
    RequestResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        request_text:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/RequestStatus'
        schedule_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        address:
          type: string
        responsible_id:
          type: string
          format: uuid
          nullable: true
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentSnapshot'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
          nullable: true
        updated_by:
          type: string
          format: uuid
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
    RequestListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RequestResponse'
    RequestUpdateRequest:
      type: object
      description: Partial update of a request. All fields are optional.
      properties:
        request_text:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/RequestStatus'
        schedule_time:
          type: string
          format: date-time
          nullable: true
        address:
          type: string
          nullable: true
        responsible_id:
          type: string
          format: uuid
          nullable: true
    AssignResponsibleRequest:
      type: object
      required: [responsible_id]
      properties:
        responsible_id:
          type: string
          format: uuid
    RequestStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/RequestStatus'
        comment:
          type: string
          nullable: true
    RequestStatusHistoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        request_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/RequestStatus'
        comment:
          type: string
          nullable: true
        changed_by:
          type: string
          format: uuid
          nullable: true
        changed_at:
          type: string
          format: date-time
    RequestStatusHistoryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RequestStatusHistoryResponse'
...